chore(experiments): delete old UI (#23813)
